direct-tree-eval: refactor: decouple ErgoTreeEvaluator (part 2)
