+++ sigmastate/src/main/scala/sigmastate/Values.scala
@@ -227,40 +227,6 @@ object Values {
     }
   }
 
-  /** The kind of the cost.*/
-  sealed abstract class CostKind
-
-  /** Descriptor of the simple fixed cost.
-    * @param cost  given cost of the operation */
-  case class FixedCost(cost: Int) extends CostKind
-
-  /** Cost of operation over collection of the known length.
-    * See for example [[Exists]], [[MapCollection]].
-    * @param baseCost cost of operation factored out of the loop iterations
-    * @param perChunkCost cost associated with each chunk of items
-    * @param chunkSize number of items in a chunk
-    */
-  case class PerItemCost(baseCost: Int, perChunkCost: Int, chunkSize: Int) extends CostKind {
-    /** Compute number of chunks necessary to cover the given number of items. */
-    def chunks(nItems: Int) = (nItems - 1) / chunkSize + 1
-
-    /** Computes the cost for the given number of items. */
-    def cost (nItems: Int): Int = {
-      val nChunks = chunks(nItems)
-      Math.addExact(baseCost, Math.multiplyExact(perChunkCost, nChunks))
-    }
-  }
-
-  /** Descriptor of the cost which depends on type. */
-  abstract class TypeBasedCost extends CostKind {
-    /** Returns cost value depending on the given type. */
-    def costFunc(tpe: SType): Int
-  }
-
-  /** Cost of operation cannot be described using fixed set of parameters.
-    * See [[EQ]], [[NEQ]]. */
-  case object DynamicCost extends CostKind
-
   /** Base class for all companion objects which are used as operation descriptors. */
   trait ValueCompanion extends SigmaNodeCompanion {
     import ValueCompanion._
